hub/requirements.txt
crewai>=0.80.0
crewai-tools>=0.17.0
langchain-openai>=0.3.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0
